brain_tumor_pipeline/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
opencv-python
Pillow
torch
torchvision
timm
tqdm

# Optional (pipeline will skip XGBoost if not installed)
xgboost

